packages/mycustomersegmentation/mycustomersegmentation-0.1.0.tar.gz/mycustomersegmentation-0.1.0/tests/test_mycustomersegmentation.py
import pytest
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
from sklearn.cluster import KMeans
from sklearn.preprocessing import StandardScaler
from sklearn.impute import SimpleImputer
from mycustomersegmentation.mycustomersegmentation import (
    load_data,
    find_missing_values,
    perform_eda,
    plot_histograms,
    plot_pairplot,
    convert_to_datetime,
    customer_data_clusters,
    customer_segmentation,
    plot_cluster_scatter
)

# Test find_missing_values function
def test_find_missing_values():
    data = pd.DataFrame({"A": [1, 2, np.nan], "B": [np.nan, 5, 6]})
    missing_values_count = find_missing_values(data)
    assert isinstance(missing_values_count, pd.Series)
    assert missing_values_count["A"] == 1
    assert missing_values_count["B"] == 1

# Test perform_eda function
def test_perform_eda():
    data = pd.DataFrame({"A": [1, 2, 3], "B": [4, 5, 6]})
    perform_eda(data)
    # Add assertions for the expected behavior of the function

# Test plot_histograms function
def test_plot_histograms():
    data = pd.DataFrame({"A": [1, 2, 3], "B": [4, 5, 6]})
    columns = ["A", "B"]
    plot_histograms(data, columns)

# Test convert_to_datetime function
def test_convert_to_datetime():
    data = pd.DataFrame({"Date": ["2021-01-01", "2021-02-01", "2021-03-01"]})
    column_name = "Date"
    converted_data = convert_to_datetime(data, column_name)
    assert isinstance(converted_data, pd.DataFrame)
    assert "date" in converted_data.columns

# Test customer_data_clusters function
def test_customer_data_clusters():
    data = pd.DataFrame({"A": [1, 2, 3], "B": [4, 5, 6]})
    num_clusters = 2
    cluster_data = customer_data_clusters(data, num_clusters)
    assert isinstance(cluster_data, pd.DataFrame)

# Test customer_segmentation function
def test_customer_segmentation():
    data = pd.DataFrame({"A": [1, 2, 3], "B": [4, 5, 6]})
    num_clusters = 2
    cluster_stats = customer_segmentation(data, num_clusters)
    assert isinstance(cluster_stats, pd.DataFrame)
